+++ tests/system/test_edb_configuration_2p0.py
@@ -484,6 +484,7 @@ class TestClass(BaseTestClass):
         }
 
         assert edbapp.configuration.load(data, apply_file=True)
+        edbapp.components.export_bom("e://temp//bom.csv")
         assert len(edbapp.components.nport_comp_definition) == 2
         assert edbapp.components.nport_comp_definition["CAPC3216X180X55ML20T25"].reference_file
         assert len(edbapp.components.nport_comp_definition["CAPC3216X180X55ML20T25"].components) == 9
